+++ src/main/java/com/configcat/ConfigurationProvider.java
@@ -1,6 +1,7 @@
 package com.configcat;
 
 import com.configcat.evaluation.EvaluationDetails;
+import com.configcat.hooks.ConfigCatHooks;
 
 import java.io.Closeable;
 import java.util.Collection;
