+++ src/test/java/com/configcat/ConfigFetcherTest.java
@@ -1,6 +1,7 @@
 package com.configcat;
 
 import com.configcat.cache.ConfigCache;
+import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
 import com.configcat.polling.PollingModes;
