+++ src/test/java/com/configcat/LazyLoadingPolicyTest.java
@@ -4,6 +4,7 @@ import com.configcat.cache.ConfigCache;
 import com.configcat.cache.FailingCache;
 import com.configcat.cache.NullConfigCache;
 import com.configcat.cache.SingleValueCache;
+import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.polling.LazyLoadingMode;
 import com.configcat.polling.PollingMode;
