+++ src/test/java/com/configcat/ManualPollingPolicyTest.java
@@ -3,6 +3,7 @@ package com.configcat;
 import com.configcat.cache.FailingCache;
 import com.configcat.cache.InMemoryCache;
 import com.configcat.cache.NullConfigCache;
+import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.polling.PollingMode;
 import com.configcat.polling.PollingModes;
